Create some headers
