+++ src/REL/d/t/d_t_holy_water.cpp
@@ -1,23 +1,5 @@
 #include <d/d_player.h>
-#include <d/tg/d_tg.h>
-#include <toBeSorted/dowsing_target.h>
-
-class dTgHolyWater_c : public dTg_c {
-public:
-    dTgHolyWater_c() : mDowsingTarget(this, DowsingTarget::SLOT_NONE) {}
-    virtual ~dTgHolyWater_c() {}
-
-    virtual int create() override;
-    virtual int doDelete() override;
-    virtual int actorExecute() override;
-    virtual int draw() override;
-
-    bool isLinkNearby();
-    bool dowsingActive();
-
-private:
-    DowsingTarget mDowsingTarget; // 0x0FC
-};
+#include <d/tg/d_t_holy_water.h>
 
 SPECIAL_ACTOR_PROFILE(TAG_HOLY_WATER, dTgHolyWater_c, fProfile::TAG_HOLY_WATER, 0x0224, 0, 4);
